Created base component.
