+++ core/ui/prime/components/src/lib/input-date/input-date.component.html
@@ -8,6 +8,7 @@
             [placeholder]="placeholder!"
             [showTime]="showTime"
             [timeOnly]="timeOnly"
+            [disabled]="disabled" 
             (onInput)="onInputChanged($event)"
             (onSelect)="onDateSelected($event)"
             [(ngModel)]="value"
