+++ core/ui/prime/components/src/lib/input-text/input-text.component.ts
@@ -1,5 +1,6 @@
 import { Component, forwardRef, Input } from "@angular/core";
-import { ControlValueAccessor, NG_VALUE_ACCESSOR } from "@angular/forms";
+import { NG_VALUE_ACCESSOR } from "@angular/forms";
+import { InputBaseComponent } from "../input-base/input-base-component.component";
 
 @Component({
     selector: "lens-input-text", 
@@ -8,42 +9,12 @@ import { ControlValueAccessor, NG_VALUE_ACCESSOR } from "@angular/forms";
         { provide: NG_VALUE_ACCESSOR, useExisting: forwardRef(() => InputTextComponent), multi: true }
     ]
 })
-export class InputTextComponent implements ControlValueAccessor {
+export class InputTextComponent extends InputBaseComponent {
     @Input() id!: string;
     @Input() placeholder?: string;
     @Input() icon!: string;
     @Input() spinIcon: boolean = false;
     @Input() iconAlign: "right" | "left" = "left";
-    @Input() disabled: boolean = false;
-
-    private _value: any;
-    public set value(value: any) {
-        this._value = value;
-        this.onChange(value);
-    }
-
-    public get value() {
-        return this._value;
-    }
-
-    private onChange = (event: any) => {};
-    private onTouched = () => {};
-
-    public writeValue(value: any): void {
-        this.value = value;
-    }
-
-    public registerOnChange(fn: any): void {
-        this.onChange = fn;
-    }
-
-    public registerOnTouched(fn: any): void {
-        this.onTouched = fn;
-    }
-
-    public setDisabledState?(isDisabled: boolean): void {
-        this.disabled = isDisabled;
-    }
 
     public onInputChanged($event: Event) {
         const value = ($event.target as HTMLInputElement).value;
